+++ lib/src/utils/thumbnails.dart
@@ -1,14 +1,14 @@
 import 'dart:typed_data';
 
 import 'package:flutter/material.dart';
-import 'package:video_editor/domain/bloc/controller.dart';
-import 'package:video_editor/domain/entities/cover_data.dart';
-import 'package:video_thumbnail/video_thumbnail.dart';
+import 'package:get_thumbnail_video/index.dart';
+import 'package:get_thumbnail_video/video_thumbnail.dart';
+import 'package:video_editor/src/controller.dart';
+import 'package:video_editor/src/models/cover_data.dart';
 
 Stream<List<Uint8List>> generateTrimThumbnails(
   VideoEditorController controller, {
   required int quantity,
-  int quality = 10,
 }) async* {
   final String path = controller.file.path;
   final double eachPart = controller.videoDuration.inMilliseconds / quantity;
@@ -16,15 +16,13 @@ Stream<List<Uint8List>> generateTrimThumbnails(
 
   for (int i = 1; i <= quantity; i++) {
     try {
-      final Uint8List? bytes = await VideoThumbnail.thumbnailData(
+      final bytes = await VideoThumbnail.thumbnailData(
         imageFormat: ImageFormat.JPEG,
         video: path,
         timeMs: (eachPart * i).toInt(),
-        quality: quality,
+        quality: controller.trimThumbnailsQuality,
       );
-      if (bytes != null) {
-        byteList.add(bytes);
-      }
+      byteList.add(bytes);
     } catch (e) {
       debugPrint(e.toString());
     }
@@ -36,7 +34,6 @@ Stream<List<Uint8List>> generateTrimThumbnails(
 Stream<List<CoverData>> generateCoverThumbnails(
   VideoEditorController controller, {
   required int quantity,
-  int quality = 10,
 }) async* {
   final int duration = controller.isTrimmed
       ? controller.trimmedDuration.inMilliseconds
@@ -46,17 +43,17 @@ Stream<List<CoverData>> generateCoverThumbnails(
 
   for (int i = 0; i < quantity; i++) {
     try {
-      final CoverData bytes = await generateSingleCoverThumbnail(
+      final CoverData? bytes = await generateSingleCoverThumbnail(
         controller.file.path,
         timeMs: (controller.isTrimmed
                 ? (eachPart * i) + controller.startTrim.inMilliseconds
                 : (eachPart * i))
             .toInt(),
-        quality: quality,
+        quality: controller.coverThumbnailsQuality,
       );
 
-      if (bytes.thumbData != null) {
-        byteList.add(bytes);
+      if (bytes?.thumbData != null) {
+        byteList.add(bytes!);
       }
     } catch (e) {
       debugPrint(e.toString());
@@ -68,18 +65,25 @@ Stream<List<CoverData>> generateCoverThumbnails(
 
 /// Generate a cover at [timeMs] in video
 ///
-/// return [CoverData] depending on [timeMs] milliseconds
-Future<CoverData> generateSingleCoverThumbnail(
+/// Returns a [CoverData] depending on [timeMs] milliseconds
+Future<CoverData?> generateSingleCoverThumbnail(
   String filePath, {
   int timeMs = 0,
   int quality = 10,
 }) async {
-  final Uint8List? thumbData = await VideoThumbnail.thumbnailData(
-    imageFormat: ImageFormat.JPEG,
-    video: filePath,
-    timeMs: timeMs,
-    quality: quality,
-  );
+  try {
+    
+    final thumbData = await VideoThumbnail.thumbnailData(
+      imageFormat: ImageFormat.JPEG,
+      video: filePath,
+      timeMs: timeMs,
+      quality: quality,
+    );
 
-  return CoverData(thumbData: thumbData, timeMs: timeMs);
+    return CoverData(thumbData: thumbData, timeMs: timeMs);
+  } catch (e) {
+    debugPrint(e.toString());
+    return null;
+  }
+
 }