+++ lib/src/widgets/crop/crop_grid.dart
@@ -1,11 +1,10 @@
 import 'dart:math';
 
 import 'package:flutter/material.dart';
-import 'package:video_editor/domain/entities/transform_data.dart';
-import 'package:video_editor/domain/helpers.dart';
-import 'package:video_editor/domain/bloc/controller.dart';
-import 'package:video_editor/ui/crop/crop_mixin.dart';
-import 'package:video_editor/ui/video_viewer.dart';
+import 'package:video_editor/src/controller.dart';
+import 'package:video_editor/src/utils/helpers.dart';
+import 'package:video_editor/src/models/transform_data.dart';
+import 'package:video_editor/src/widgets/crop/crop_mixin.dart';
 
 @protected
 enum CropBoundaries {
@@ -304,7 +303,9 @@ class _CropGridViewerState extends State<CropGridViewer> with CropPreviewMixin {
     // don't apply changes if out of bounds
     if (newRect.width < minRectSize ||
         newRect.height < minRectSize ||
-        !isRectContained(layout, newRect)) return;
+        !isRectContained(layout, newRect)) { 
+          return;
+    }
 
     rect.value = newRect;
   }